convert admin section to hooks and ts
